+++ lib/src/constants.dart
@@ -48,6 +48,7 @@ class Constants {
   static const String SET_PUSH_TOKEN = 'setPushToken';
   static const String DISABLE_IN_APP_MESSAGES = 'disableInAppMessages';
   static const String ENABLE_IN_APP_MESSAGES = 'enableInAppMessages';
+  static const String HANDLE_UNIVERSAL_LINK = 'handleUniversalLink';
 
   static const String ENABLE_CARRIER_COLLECTION = 'enableCarrierCollection';
   static const String ENABLE_IP_COLLECTION = 'enableIpCollection';